requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
audioread==2.1.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.2
ffmpeg-python==0.2.0
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.36.1
h11==0.12.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpcore==0.12.3
httpx==0.16.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.7.2
joblib==1.0.1
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.4
norbert==0.2.1
numba==0.52.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.2
Pillow==8.1.0
pooch==1.3.0
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.4.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
sniffio==1.2.0
SoundFile==0.10.3.post1
spleeter==2.2.1
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
